chore: remove fun stuff (#656)
